Split out sprite rendering
