+++ crates/bevy_internal/src/default_plugins.rs
@@ -42,6 +42,8 @@ plugin_group! {
         bevy_anti_aliasing:::AntiAliasingPlugin,
         #[cfg(feature = "bevy_sprite")]
         bevy_sprite:::SpritePlugin,
+        #[cfg(feature = "bevy_sprite_render")]
+        bevy_sprite_render:::SpriteRenderingPlugin,
         #[cfg(feature = "bevy_text")]
         bevy_text:::TextPlugin,
         #[cfg(feature = "bevy_ui")]
