+++ crates/bevy_internal/src/prelude.rs
@@ -68,6 +68,10 @@ pub use crate::scene::prelude::*;
 pub use crate::sprite::prelude::*;
 
 #[doc(hidden)]
+#[cfg(feature = "bevy_sprite_render")]
+pub use crate::sprite_render::prelude::*;
+
+#[doc(hidden)]
 #[cfg(feature = "bevy_text")]
 pub use crate::text::prelude::*;
