--- crates/bevy_sprite/src/mesh2d/mesh2d_view_types.wgsl
@@ -1,4 +1,0 @@
-#define_import_path bevy_sprite::mesh2d_view_types
-
-#import bevy_render::view
-#import bevy_render::globals
