+++ crates/bevy_sprite/src/texture_slice/mod.rs
@@ -1,14 +1,9 @@
 mod border_rect;
-mod computed_slices;
 mod slicer;
 
 use bevy_math::{Rect, Vec2};
 pub use border_rect::BorderRect;
 pub use slicer::{SliceScaleMode, TextureSlicer};
-
-pub(crate) use computed_slices::{
-    compute_slices_on_asset_event, compute_slices_on_sprite_change, ComputedTextureSlices,
-};
 
 /// Single texture slice, representing a texture rect to draw in a given area
 #[derive(Debug, Clone, PartialEq)]
