+++ crates/bevy_sprite_render/Cargo.toml
@@ -1,50 +1,49 @@
 [package]
-name = "bevy_ui_render"
+name = "bevy_sprite_render"
 version = "0.17.0-dev"
 edition = "2024"
-description = "Provides rendering functionality for Bevy UI"
-homepage = "https://bevyengine.org"
+description = "Provides sprite rendering functionality for Bevy Engine"
+homepage = "https://bevy.org"
 repository = "https://github.com/bevyengine/bevy"
 license = "MIT OR Apache-2.0"
 keywords = ["bevy"]
+
+[features]
+webgl = []
+webgpu = []
 
 [dependencies]
 # bevy
 bevy_app = { path = "../bevy_app", version = "0.17.0-dev" }
 bevy_asset = { path = "../bevy_asset", version = "0.17.0-dev" }
-bevy_camera = { path = "../bevy_camera", version = "0.17.0-dev" }
 bevy_color = { path = "../bevy_color", version = "0.17.0-dev" }
 bevy_core_pipeline = { path = "../bevy_core_pipeline", version = "0.17.0-dev" }
-bevy_derive = { path = "../bevy_derive", version = "0.17.0-dev" }
 bevy_ecs = { path = "../bevy_ecs", version = "0.17.0-dev" }
 bevy_image = { path = "../bevy_image", version = "0.17.0-dev" }
-bevy_math = { path = "../bevy_math", version = "0.17.0-dev" }
+bevy_camera = { path = "../bevy_camera", version = "0.17.0-dev" }
 bevy_mesh = { path = "../bevy_mesh", version = "0.17.0-dev" }
-bevy_reflect = { path = "../bevy_reflect", version = "0.17.0-dev" }
+bevy_math = { path = "../bevy_math", version = "0.17.0-dev" }
 bevy_shader = { path = "../bevy_shader", version = "0.17.0-dev" }
-bevy_render = { path = "../bevy_render", version = "0.17.0-dev" }
 bevy_sprite = { path = "../bevy_sprite", version = "0.17.0-dev" }
-bevy_sprite_render = { path = "../bevy_sprite_render", version = "0.17.0-dev" }
-bevy_picking = { path = "../bevy_picking", version = "0.17.0-dev", optional = true }
+# bevy_picking = { path = "../bevy_picking", version = "0.17.0-dev", optional = true }
+bevy_reflect = { path = "../bevy_reflect", version = "0.17.0-dev" }
+bevy_render = { path = "../bevy_render", version = "0.17.0-dev" }
 bevy_transform = { path = "../bevy_transform", version = "0.17.0-dev" }
 bevy_utils = { path = "../bevy_utils", version = "0.17.0-dev" }
+# bevy_window = { path = "../bevy_window", version = "0.17.0-dev", optional = true }
+bevy_derive = { path = "../bevy_derive", version = "0.17.0-dev" }
 bevy_platform = { path = "../bevy_platform", version = "0.17.0-dev", default-features = false, features = [
   "std",
 ] }
-bevy_ui = { path = "../bevy_ui", version = "0.17.0-dev" }
-bevy_text = { path = "../bevy_text", version = "0.17.0-dev", default-features = false }
-bevy_window = { path = "../bevy_window", version = "0.17.0-dev", default-features = false }
 
 # other
-bytemuck = { version = "1.5", features = ["derive"] }
-derive_more = { version = "1", default-features = false, features = ["from"] }
+bytemuck = { version = "1", features = ["derive", "must_cast"] }
+fixedbitset = "0.5"
+derive_more = { version = "2", default-features = false, features = ["from"] }
+bitflags = "2.3"
+# radsort = "0.1"
+nonmax = "0.5"
 tracing = { version = "0.1", default-features = false, features = ["std"] }
-
-[features]
-default = []
-serialize = ["bevy_math/serialize", "bevy_platform/serialize"]
-bevy_ui_picking_backend = ["bevy_picking"]
-bevy_ui_debug = []
 
 [lints]
 workspace = true
