+++ crates/bevy_sprite_render/README.md
@@ -1,0 +1,7 @@
+# Bevy Sprite
+
+[![License](https://img.shields.io/badge/license-MIT%2FApache-blue.svg)](https://github.com/bevyengine/bevy#license)
+[![Crates.io](https://img.shields.io/crates/v/bevy_sprite_render.svg)](https://crates.io/crates/bevy_sprite_render)
+[![Downloads](https://img.shields.io/crates/d/bevy_sprite_render.svg)](https://crates.io/crates/bevy_sprite_render)
+[![Docs](https://docs.rs/bevy_sprite_render/badge.svg)](https://docs.rs/bevy_sprite_render/latest/bevy_sprite_render/)
+[![Discord](https://img.shields.io/discord/691052431525675048.svg?label=&logo=discord&logoColor=ffffff&color=7389D8&labelColor=6A7EC2)](https://discord.gg/bevy)
