+++ crates/bevy_sprite_render/src/mesh2d/color_material.wgsl
@@ -1,4 +1,4 @@
-#import bevy_sprite::{
+#import bevy_sprite_render::{
     mesh2d_vertex_output::VertexOutput,
     mesh2d_view_bindings::view,
 }
