+++ crates/bevy_sprite_render/src/mesh2d/mesh2d_bindings.wgsl
@@ -1,6 +1,6 @@
-#define_import_path bevy_sprite::mesh2d_bindings
+#define_import_path bevy_sprite_render::mesh2d_bindings
 
-#import bevy_sprite::mesh2d_types::Mesh2d
+#import bevy_sprite_render::mesh2d_types::Mesh2d
 
 #ifdef PER_OBJECT_BUFFER_BATCH_SIZE
 @group(1) @binding(0) var<uniform> mesh: array<Mesh2d, #{PER_OBJECT_BUFFER_BATCH_SIZE}u>;
