+++ crates/bevy_sprite_render/src/mesh2d/mesh2d_view_bindings.wgsl
@@ -1,4 +1,4 @@
-#define_import_path bevy_sprite::mesh2d_view_bindings
+#define_import_path bevy_sprite_render::mesh2d_view_bindings
 
 #import bevy_render::view::View
 #import bevy_render::globals::Globals
