+++ crates/bevy_sprite_render/src/mesh2d/mesh2d_view_types.wgsl
@@ -1,0 +1,4 @@
+#define_import_path bevy_sprite_render::mesh2d_view_types
+
+#import bevy_render::view
+#import bevy_render::globals
