+++ crates/bevy_sprite_render/src/mesh2d/wireframe2d.wgsl
@@ -1,4 +1,4 @@
-#import bevy_sprite::mesh2d_vertex_output::VertexOutput
+#import bevy_sprite_render::mesh2d_vertex_output::VertexOutput
 
 struct PushConstants {
     color: vec4<f32>
