+++ crates/bevy_sprite_render/src/render/sprite.wgsl
@@ -7,7 +7,7 @@
     view::View,
 }
 
-#import bevy_sprite::sprite_view_bindings::view
+#import bevy_sprite_render::sprite_view_bindings::view
 
 struct VertexInput {
     @builtin(vertex_index) index: u32,
