+++ crates/bevy_sprite_render/src/sprite.rs
@@ -1,0 +1,7 @@
+
+use bevy_render::sync_world::SyncToRenderWorld;
+
+// TODO: Find a way to:
+
+// #[require(SyncToRenderWorld)]
+// pub struct Sprite;
