+++ crates/bevy_sprite_render/src/texture_slice/mod.rs
@@ -1,0 +1,5 @@
+mod computed_slices;
+
+pub(crate) use computed_slices::{
+    compute_slices_on_asset_event, compute_slices_on_sprite_change, ComputedTextureSlices,
+};
