+++ crates/bevy_sprite_render/src/tilemap_chunk/tilemap_chunk_material.rs
