+++ crates/bevy_sprite_render/src/tilemap_chunk/tilemap_chunk_material.wgsl
@@ -1,4 +1,4 @@
-#import bevy_sprite::{
+#import bevy_sprite_render::{
     mesh2d_functions as mesh_functions,
     mesh2d_view_bindings::view,
     mesh2d_vertex_output::VertexOutput,
