+++ crates/bevy_ui_render/src/ui_texture_slice_pipeline.rs
@@ -25,7 +25,8 @@ use bevy_render::{
 };
 use bevy_render::{sync_world::MainEntity, RenderStartup};
 use bevy_shader::Shader;
-use bevy_sprite::{SliceScaleMode, SpriteAssetEvents, SpriteImageMode, TextureSlicer};
+use bevy_sprite::{SliceScaleMode, SpriteImageMode, TextureSlicer};
+use bevy_sprite_render::SpriteAssetEvents;
 use bevy_ui::widget;
 use bevy_utils::default;
 use binding_types::{sampler, texture_2d};
