+++ examples/2d/tilemap_chunk.rs
@@ -2,7 +2,7 @@
 
 use bevy::{
     prelude::*,
-    sprite::{TileData, TilemapChunk, TilemapChunkTileData},
+    sprite_rendering::{TileData, TilemapChunk, TilemapChunkTileData},
 };
 use rand::{Rng, SeedableRng};
 use rand_chacha::ChaCha8Rng;
